move JobContext to context module
